Fix cem_mpc types

+++ src/gui_application/math.cpp
@@ -3,14 +3,14 @@
 namespace RoboticsSandbox
 {
 
-float angle_diff(float a, float b)
+double angle_diff(double a, double b)
 {
-  float dif = fmod(b - a + 180.0f, 360.0f);
+  double dif = fmod(b - a + 180.0f, 360.0f);
   if (dif < 0)
   {
-    dif += 360.0f;
+    dif += 360.0;
   }
-  return dif - 180.0f;
+  return dif - 180.0;
 }
 
 } // namespace RoboticsSandbox
